Cleaned up packages, removed scraper files, initial work on profile page
