+++ client/package.json
@@ -6,7 +6,18 @@
     "axios": "^0.19.0",
     "react": "^16.11.0",
     "react-dom": "^16.11.0",
-    "react-scripts": "^3.3.0"
+    "react-scripts": "^3.3.0",
+    "@material-ui/core": "^4.9.5",
+    "@material-ui/icons": "^4.5.1",
+    "@material-ui/lab": "^4.0.0-alpha.45",
+    "prop-types": "^15.7.2",
+    "react-image-lightbox": "^5.1.1",
+    "react-redux": "^7.2.0",
+    "react-router-dom": "^5.1.0",
+    "redux": "^4.0.5",
+    "redux-devtools-extension": "^2.13.8",
+    "redux-persist": "^6.0.0",
+    "redux-thunk": "^2.3.0"
   },
   "scripts": {
     "start": "react-scripts start",
