+++ client/src/components/Layout/IconHeader.js
@@ -2,6 +2,7 @@ import * as colors from '../../helpers/colors';
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import { Typography, Grid } from '@material-ui/core';
+import { MiniDivider } from './MiniDivider';
 
 const useStyles = makeStyles(theme => ({
   paper: {
@@ -11,10 +12,7 @@ const useStyles = makeStyles(theme => ({
     flexDirection: 'column',
     alignItems: 'center'
   },
-  divider: {
-    width: '40px',
-    border: 0
-  },
+
   paperHeader: {
     fontFamily: 'Raleway',
     textTransform: 'uppercase'
@@ -42,20 +40,14 @@ export const IconHeader = ({ icon, text, subheader }) => {
         </Grid>
         <Grid item>
           <Typography
-            variant='h4'
+            variant={subheader ? 'h4' : 'h3'}
             className={classes.paperHeader}
             style={subheader ? { color: 'black' } : { color: color }}>
             {text}
           </Typography>
         </Grid>
       </Grid>
-      <hr
-        className={classes.divider}
-        style={{
-          borderTop: '2px solid ' + color,
-          marginBottom: subheader ? 40 : 0
-        }}
-      />
+      <MiniDivider color={color} />
     </div>
   );
 };
