+++ package.json
@@ -21,23 +21,15 @@
   "homepage": "https://github.com/Foxyf76/Automater-POC-javascript#readme",
   "dependencies": {
     "@google-cloud/vision": "^1.8.0",
-    "@material-ui/core": "^4.9.5",
-    "@material-ui/icons": "^4.5.1",
-    "@material-ui/lab": "^4.0.0-alpha.45",
     "@teachablemachine/image": "^0.8.4",
     "@tensorflow-models/coco-ssd": "^2.0.0",
     "@tensorflow-models/mobilenet": "^2.0.4",
     "@tensorflow-models/toxicity": "^1.2.2",
     "@tensorflow/tfjs-node": "1.4.0",
-    "atob": "^2.1.2",
     "axios": "^0.19.0",
-    "base64-arraybuffer": "^0.2.0",
-    "base64-img": "^1.0.4",
-    "base64-to-uint8array": "^1.0.0",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
     "canvas": "^2.6.1",
-    "chromedriver": "^77.0.0",
     "config": "^3.2.2",
     "connect-mongo": "^3.1.2",
     "cookie-parser": "^1.4.4",
@@ -50,40 +42,15 @@
     "express-validator": "^6.4.0",
     "face-api.js": "^0.22.0",
     "fbgraph": "^1.4.4",
-    "gm": "^1.23.1",
-    "gravatar": "^1.8.0",
-    "jimp": "^0.9.3",
-    "jpeg-js": "^0.3.6",
     "jsdom": "^15.2.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.7.5",
     "node-fetch": "^2.6.0",
     "passport": "^0.4.0",
     "passport-facebook": "^3.0.0",
-    "passport-jwt": "^4.0.0",
-    "passport-local": "^1.0.0",
     "passport-twitter": "^1.0.4",
-    "prop-types": "^15.7.2",
-    "puppeteer": "^2.0.0",
-    "python-shell": "^1.0.8",
-    "react-bounding-box": "^0.5.9",
-    "react-image-lightbox": "^5.1.1",
-    "react-images": "^1.1.0-beta.3",
-    "react-imgpro": "^1.4.1",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.1.0",
-    "react-spinners-css": "^1.1.7",
-    "react-window": "^1.8.5",
-    "redux": "^4.0.5",
-    "redux-devtools-extension": "^2.13.8",
-    "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.3.0",
     "request": "^2.88.0",
-    "selenium-webdriver": "^4.0.0-alpha.5",
     "sharp": "^0.24.0",
-    "tesseract.js": "^2.0.2",
-    "twit": "^2.2.11",
-    "twitter": "^1.7.1",
     "twitter-lite": "^0.9.4",
     "uuid": "^7.0.1"
   },
