+++ server.js
@@ -50,7 +50,6 @@ app.get('/', (req, res) => {
 
 // Define routes here
 app.use('/api/auth', require('./routes/api/auth'));
-app.use('/api/scrape', require('./routes/api/scrape'));
 app.use('/api/passport-auth', require('./routes/api/auth-passport'));
 app.use('/api/classifier', require('./routes/api/classifier'));
